Fix lat & lon types

+++ lib/PHPAPI/src/ApiResource/PositionRecordResponse.php
@@ -13,9 +13,9 @@ class PositionRecordResponse
 
     public string $datetime = '';
 
-    public int $lat = 0;
+    public float $lat = 0.0;
 
-    public int $lon = 0;
+    public float $lon = 0.0;
 
     public int $minutes = 0;
